+++ include/dtorr/metadata.h
@@ -3,7 +3,7 @@
 
 #include "structs.h"
 
-dtorr_torrent* load_torrent_metadata(dtorr_config* config, char* data, unsigned long data_len);
+dtorr_torrent* load_torrent_metadata(dtorr_config* config, char* data, unsigned long long data_len);
 
 void free_torrent(dtorr_torrent* torrent);
